Fix for app crashing on first start or recompile
Wait for elastic to start before handling requests.
